+++ src/locales/zh-Hans.ts
@@ -69,6 +69,10 @@ export default {
             listItem: "列表项",
         },
     },
+    search_description: {
+        doc_count: "文档数",
+        block_count: "块数",
+    },
     search_result: "@:search@:result",
     search_settings: "@:search@:settings",
     server: "服务",